refactor: 코드 포맷 및 뷰 로직 변경
