+++ src/app.js
@@ -10,30 +10,6 @@ import { setModeToggleButton } from "./components/main/mode-toggle-button/index.
 import { setOptionToggleButton } from "./components/main/option-toggle-button/index.js";
 
 (function init() {
-  const slider = document.querySelector(".list-view_category-bar");
-  let isDown = false;
-  let startX;
-  let scrollLeft;
-
-  slider.addEventListener("mousedown", (e) => {
-    isDown = true;
-    startX = e.pageX - slider.offsetLeft;
-    scrollLeft = slider.scrollLeft;
-  });
-  slider.addEventListener("mouseleave", () => {
-    isDown = false;
-  });
-  slider.addEventListener("mouseup", () => {
-    isDown = false;
-  });
-  slider.addEventListener("mousemove", (e) => {
-    if (!isDown) return;
-    e.preventDefault();
-    const x = e.pageX - slider.offsetLeft;
-    const walk = x - startX; //scroll-fast
-    slider.scrollLeft = scrollLeft - walk;
-  });
-
   setSubscribe();
   setHeader();
   setGrid();
