+++ src/components/common/alert/index.js
@@ -1,6 +1,6 @@
 import { setEvents, showAlert } from "./Alert.js";
-import { alertMsgState } from "../../../store/store.js";
-import { subscribe } from "../../../observer/observer.js";
+import { alertMsgState } from "../../../store/storeKey.js";
+import { subscribe } from "../../../store/observer.js";
 
 export const setAlert = () => {
   subscribe(alertMsgState, showAlert);
