+++ src/components/common/snack-bar/index.js
@@ -1,6 +1,6 @@
-import { subscribe } from "../../../observer/observer.js";
-import { snackBarMsgState } from "../../../store/store.js";
+import { subscribe } from "../../../store/observer.js";
 import { setEvents, showSnackBar } from "./SnackBar.js";
+import { snackBarMsgState } from "../../../store/storeKey.js";
 
 export const setSnackBar = () => {
   subscribe(snackBarMsgState, showSnackBar);
