+++ src/components/main/option-toggle-button/index.js
@@ -1,5 +1,5 @@
-import { subscribe } from "../../../observer/observer.js";
-import { viewOptionState } from "../../../store/store.js";
+import { subscribe } from "../../../store/observer.js";
+import { viewOptionState } from "../../../store/storeKey.js";
 import { changeActivateState, setEvents } from "./OptionToggleButton.js";
 
 export const setOptionToggleButton = () => {
