+++ src/components/main/subscribe-button/index.js
@@ -1,6 +1,6 @@
 import { customFetch } from "../../../utils/utils.js";
-import { setState } from "../../../observer/observer.js";
-import { subscribeState } from "../../../store/store.js";
+import { setState } from "../../../store/observer.js";
+import { subscribeState } from "../../../store/storeKey.js";
 
 export const setSubscribe = async () => {
   const subList = await customFetch("../../../mocks/subscribe.json");
