+++ src/components/main/view-toggle-button/index.js
@@ -1,5 +1,5 @@
-import { viewState } from "../../../store/store.js";
-import { subscribe } from "../../../observer/observer.js";
+import { viewState } from "../../../store/storeKey.js";
+import { subscribe } from "../../../store/observer.js";
 import { setCurrentView, setEvents } from "./ViewToggleButton.js";
 
 export const setViewToggleButton = () => {
